+++ CHANGES.md
@@ -6,5 +6,7 @@
 * `--filter` now only works on named examples
 * Added the `--flagged-only` option
 * Added the `--format` option and the json format
+* Added the `--runner` option
 * Added the `--ignore-unknown-annotations` option
 * Annotation arguments can now be enclosed in single as well as double qoutes
+* Included examples are no longer wrapped in output buffering